+++ old/src/applySettingsHook.ts
