+++ old/src/fpsCounter.ts
