+++ old/src/friendSearch.ts
