+++ old/src/fullscreenHook.ts
