+++ old/src/hashManager.ts
