+++ old/src/mapIdentifier.ts
