+++ old/src/matchEndHook.ts
