+++ old/src/preloaderHook.ts
