+++ old/src/updateChecker.ts
