+++ old/src/userCommunicationProtocol.ts
